Add CollectionBuilder for ReadOnlyValueSet<T>
